+++ src/hooks/index.ts
@@ -1,4 +1,5 @@
 import useOrientationChange from './useOrientationChange';
 import useAndroidKeyboard from './useAndroidKeyboard';
+import useScrollTop from './useScrollTop';
 
-export { useOrientationChange, useAndroidKeyboard };
+export { useOrientationChange, useAndroidKeyboard, useScrollTop };