+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+#### v2.1.0
+- [+] useAndroidKeyboard
+- [+] useOrientationChange
+- [+] useScrollTop
+- [*] fix JSDoc syntax
+- [*] checkMobile adds classname 'desktop'
+
 #### v2.0.6
 - [+] markup: optional parameter withCheckMobile
 - [*] markup: parameter markupConst is now optional